Move Mongojack Micronaut integration to own module and use compiletime bean introspection for modules beside backend
