--- backend/gradle.properties
@@ -1,1 +1,0 @@
-micronautVersion=3.2.0
