+++ dto_common/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     java
+    id("io.micronaut.library") version "3.0.1"
 }
 
 group = "club.devcord.devmarkt"
@@ -16,17 +17,10 @@ java {
 }
 
 dependencies {
+    annotationProcessor("io.micronaut:micronaut-inject-java")
+
     compileOnly("org.mongojack:mongojack:4.3.0")
-
-    implementation("io.swagger.core.v3:swagger-annotations:2.1.11")
-
+    compileOnly("io.swagger.core.v3:swagger-annotations:2.1.11")
     testImplementation("org.junit.jupiter:junit-jupiter-api:5.8.2")
     testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine")
-}
-
-tasks.getByName<Test>("test") {
-    useJUnitPlatform()
-    this.testLogging {
-        this.showStandardStreams = true
-    }
 }
