+++ micronaut-mongojack/src/main/java/club/devcord/devmarkt/micronaut_mongojack/NoCollectionTypeProvided.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package club.devcord.devmarkt.mongodb;
+package club.devcord.devmarkt.micronaut_mongojack;
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
+public class NoCollectionTypeProvided extends RuntimeException {
 
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Collection {
-  Class<?> value();
+  public NoCollectionTypeProvided() {
+    super("No collection type provided");
+  }
+
 }
