+++ settings.gradle.kts
@@ -1,3 +1,4 @@
 rootProject.name = "devmarkt-backend"
 include("backend")
 include("dto_common")
+include("micronaut-mongojack")
